+++ src/main/java/org/pr/project/strategies/NumericFilteringStrategy.java
@@ -1,8 +1,6 @@
 package org.pr.project.strategies;
 
-import java.math.BigDecimal;
-
-public interface NumericFilteringStrategy extends FilteringStrategy<BigDecimal>{
+public interface NumericFilteringStrategy extends FilteringStrategy<Double>{
 
 
 }
